requirements.txt
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==9.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyQt5==5.15.11
PyQt5_Frameless_Window==0.4.3
PyQt5_sip==12.15.0
PyQt_Fluent_Widgets==1.7.1
pywin32==302
ffmpeg-python==0.2.0